ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.196
openai==1.3.5
python-jose==3.3.0
redis==5.0.1
Pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0
